+++ stapi-pydantic/src/stapi_pydantic/constants.py
@@ -1,0 +1,2 @@
+STAPI_VERSION = "0.1.0"
+"""The default STAPI version for this library."""