+++ stapi-pydantic/src/stapi_pydantic/order.py
@@ -13,6 +13,7 @@ from pydantic import (
     field_validator,
 )
 
+from .constants import STAPI_VERSION
 from .datetime_interval import DatetimeInterval
 from .filter import CQL2Filter
 from .opportunity import OpportunityProperties
@@ -35,13 +36,13 @@ class OrderStatusCode(StrEnum):
     accepted = "accepted"
     rejected = "rejected"
     completed = "completed"
-    canceled = "canceled"
+    cancelled = "cancelled"
     scheduled = "scheduled"
     held = "held"
     processing = "processing"
     reserved = "reserved"
     tasked = "tasked"
-    user_canceled = "user_canceled"
+    user_cancelled = "user_cancelled"
 
 
 class OrderStatus(BaseModel):
@@ -54,7 +55,10 @@ class OrderStatus(BaseModel):
     model_config = ConfigDict(extra="allow")
 
 
-class OrderStatuses[T: OrderStatus](BaseModel):
+T = TypeVar("T", bound=OrderStatus)
+
+
+class OrderStatuses(BaseModel, Generic[T]):
     statuses: list[T]
     links: list[Link] = Field(default_factory=list)
 
@@ -63,10 +67,10 @@ class OrderSearchParameters(BaseModel):
     datetime: DatetimeInterval
     geometry: Geometry
     # TODO: validate the CQL2 filter?
-    filter: CQL2Filter | None = None
+    filter: CQL2Filter | None = None  # type: ignore [type-arg]
 
 
-class OrderProperties[T: OrderStatus](BaseModel):
+class OrderProperties(BaseModel, Generic[T]):
     product_id: str
     created: AwareDatetime
     status: T
@@ -79,11 +83,13 @@ class OrderProperties[T: OrderStatus](BaseModel):
 
 
 # derived from geojson_pydantic.Feature
-class Order[T: OrderStatus](_GeoJsonBase):
+class Order(_GeoJsonBase, Generic[T]):
     # We need to enforce that orders have an id defined, as that is required to
     # retrieve them via the API
     id: StrictStr
     type: Literal["Feature"] = "Feature"
+    stapi_type: Literal["Order"] = "Order"
+    stapi_version: str = STAPI_VERSION
 
     geometry: Geometry = Field(...)
     properties: OrderProperties[T] = Field(...)
@@ -102,7 +108,7 @@ class Order[T: OrderStatus](_GeoJsonBase):
 
 
 # derived from geojson_pydantic.FeatureCollection
-class OrderCollection[T: OrderStatus](_GeoJsonBase):
+class OrderCollection(_GeoJsonBase, Generic[T]):
     type: Literal["FeatureCollection"] = "FeatureCollection"
     features: list[Order[T]]
     links: list[Link] = Field(default_factory=list)
@@ -124,7 +130,7 @@ class OrderPayload(BaseModel, Generic[ORP]):
     datetime: DatetimeInterval
     geometry: Geometry
     # TODO: validate the CQL2 filter?
-    filter: CQL2Filter | None = None
+    filter: CQL2Filter | None = None  # type: ignore [type-arg]
 
     order_parameters: ORP