+++ stapi-pydantic/src/stapi_pydantic/product.py
@@ -1,13 +1,12 @@
 from enum import StrEnum
-from typing import Any, Literal, Self
+from typing import Any, Literal, Self, TypeAlias
 
 from pydantic import AnyHttpUrl, BaseModel, Field
 
-from .opportunity import OpportunityProperties
-from .order import OrderParameters
+from .constants import STAPI_VERSION
 from .shared import Link
 
-type Constraints = BaseModel
+Constraints: TypeAlias = BaseModel
 
 
 class ProviderRole(StrEnum):
@@ -30,7 +29,9 @@ class Provider(BaseModel):
 
 
 class Product(BaseModel):
-    type_: Literal["Product"] = Field(default="Product", alias="type")
+    type_: Literal["Collection"] = Field(default="Collection", alias="type")
+    stapi_type: Literal["Product"] = "Product"
+    stapi_version: str = STAPI_VERSION
     conformsTo: list[str] = Field(default_factory=list)
     id: str
     title: str = ""
@@ -39,37 +40,6 @@ class Product(BaseModel):
     license: str
     providers: list[Provider] = Field(default_factory=list)
     links: list[Link] = Field(default_factory=list)
-
-    # we don't want to include these in the model fields
-    _constraints: type[Constraints]
-    _opportunity_properties: type[OpportunityProperties]
-    _order_parameters: type[OrderParameters]
-
-    def __init__(
-        self,
-        *args: Any,
-        constraints: type[Constraints],
-        opportunity_properties: type[OpportunityProperties],
-        order_parameters: type[OrderParameters],
-        **kwargs: Any,
-    ) -> None:
-        super().__init__(*args, **kwargs)
-
-        self._constraints = constraints
-        self._opportunity_properties = opportunity_properties
-        self._order_parameters = order_parameters
-
-    @property
-    def constraints(self) -> type[Constraints]:
-        return self._constraints
-
-    @property
-    def opportunity_properties(self) -> type[OpportunityProperties]:
-        return self._opportunity_properties
-
-    @property
-    def order_parameters(self) -> type[OrderParameters]:
-        return self._order_parameters
 
     def with_links(self, links: list[Link] | None = None) -> Self:
         if not links: